Move rendering -> render, hal -> render/hal
