+++ render/sse/main.go
